Added training
The neural net can now be trained, however there are two versions: 1 with global variables that print progress and a function oriented approach that does not. This is because I failed in figuring out how to add progress printing to a function oriented approach.  

next TODO: loss function graphs + validation