Finished initial NN build
1. Neural Net now trains
2. Plots the loss, grid data and neural net predicted data on one plot
3. added function to collect headers
Intend to make inputting the inputs and outputs for the neural net more versatile using the header collecting function.
Further testing is required on training the NN because with a architecture of 100 nodes per layer and 6 hidden layers, trained for 100,000 epochs (which is certainly a sufficient number of epochs as both the MAE and MSE are flat) the NN is under fitting. Hopefully this can be improved by relieving the age limits and by experimenting with the regularization. 
next TODO: add variables used to train NN to graph