Grid NN: neuralStellar functions to be called during project (basically just a massive cleanup/reorganization of code)
Decided to isolate out the functions I have written so far and pour them all into objects and definitions as a nice little package that any of us can call from later on. This tackles everything (so far) that is needed surrounding loading in grids, running NNs, evaluating and predicting with NNs, plotting history and HR diagrams, bla bla bla. Next to do is to write a proper doc string for it.