sorted issue with parameter order
Issue occurred due to the plotting of the results being hindered by there not being a way of telling the order of the parameters in the output_data. This should now be fixed. Also changed some function parameters as they become redundant.
However new issue with plotting has also occurred.