Grid NN: grid03 and training of small grid using 4 inputs and 3 outputs
grid03 is made after neuralStellar is upgraded to take in more than just mass and age as inputs, but also Y, feh and MLT if available, and outputs delta nu on top of L and Teff. This version shows the combination of NN settings that gave the lowest MAE loss seen given the same number of epochs trained.