History file now works, improved graphs, dex calcs
Currently the preferences file is loaded with yaml but the history file is saved using dill, but it saves the number of epochs that have been run, MAE, val_MAE and current weights+biases of the neural net. It also updates so you can re-run the file and know exactly how many epochs the neural net has been trained for. 
Improved results graph which now includes the newly calculated dex values, the colourbar for the HR diagrams is now more obvious and has descriptive text.