Artificial open cluster HBM test for NN from trial28
A HBM test as described by Guy in meeting 04/12/19:
Construct an open cluster with stars of similar ages from the grid, and get the HBM to model the population as any real open cluster. This would test the NN's accuracy and how effective can HBM use the NN to find global minimums in the posterior of the whole function.
stars are taken with age 4.5Gyr plus or minus 0.225Gyr (329 stars match)
Observed data are taken as the observables in the selected grid data (Teff, L, delnu).
Observational error: typical observational error for stars.
Hyper-prior: mean age = N(4.5,0.1), spread age = halfnormal(0.1)
Mass input to the NN is generated as a flat uniform function from 0.8 to 1.2 Msun (limits of small_grid data), for 329 stars
feh and MLT are kept as constant (0 and 1.9) as they never varied in the small_grid.

Results:
Horrible fit and I have no idea how to fix it. I tried many different distributions in generating the NN inputs (namely age and mass) but none seem to return a sensible Rhat value for any of the estimated quantities. Temporarily given up on this task.