Adopted things about HBM learnt from testing HBM on NN generated data only, and used them back on this problem
Results are that the bias in age still exists, but the width of the bias shrunk. The source of the bias is also identified, which is from a mixed contribution of how MS stars don't vary much in observables when ageing, and its effect being compounded by the non-perfect NN approximation of the grid at the hook regions of isochrones.