The results from testing single track pipeline
After training a large number of neural networks on just single tracks of just the RGB, a somewhat optimal training pipeline was found which produced good losses and HR diagrams where the red clump was a good shape in the right place. 
This file shows the consistency of this pipeline when used on all 25 RGB tracks in the standard training grid.