First training on just MS and Subgiant
Since yesterday we've changed the cut we use to separate the subgaint and RGB stars so this training is now obsolete but does show that we still get good results using the old pipeline. However, we would like to get down the 6 kelvin median error and the extended isochrone shows that the neural net will make predictions outside of its training region that cross back over its values in the trained regions which would make HBM sampling difficult. We hope the new cut which keeps it the lower section of the RGB will prevent predictions crossing back over themselves.