New file to determine cluster members
step_1_create_files is a function which returns a csv containing only the useful data about all of the suspected stars in the cluster regardless of membership probability, as well as a txt file containing the GAIA DR2 IDs for use with the SIMBAD and GAIA query tools to get effective temperatures and luminosities (the query itself is done by some of Hin's code).

step2_membership uses the csv created in step1 (unless the open cluster being considered is M67 in which it uses the other paper with more accurate membership probabilities because you get 4 times as many stars) and a file returned from Hin's querying, which allows for the membership probabilities to be matched to the correct star's effective temperatures and luminosities. The largest number of stars where the uncertainty per star is 1% is determined and those stars and their useful data are written to a new csv. 
There is an additional luminosity calculation done using apparent V-band mag, Teff and a distance from some paper (we haven't yet figured out using parallax as a distance measurement).