plots and results for NGC188 trial3 using NN 146
Good Rhats and sampling results, but isochrone does not entirely agree with the data. Most likely caused by the lack of datapoints (9 stars) and all of them in the sub-giant branch, which is more problematic in NN fitting than in MS. Still going ahead and use this as our final result but will address shortcomings on this cluster.