plots and results for NGC6819 trial2 with NN 146
Problem: the HBM is unable to estimate NGC6819 effectively due to no isochrone really matching the distribution of stars particularly, this is seen in the lower plots of the notebook, showing there is no stars in the hook region of the isochrone derived from literature values and assuming a mixing length = 1.9. There is also a bunch of unexpected stars at the inside of the hook region, which could be throwing the HBM off. We decided to test removing those inside stars and putting back in some stars that we removed, that used to belong around the hook region for the next trial.